Unify FakeViewModel with MockViewModel
